requirements.txt
# Core ML and NLP
spacy>=3.6.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
tensorflow>=2.13.0
torch>=2.0.0
transformers>=4.21.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.0

# Time Series Analysis
statsmodels>=0.14.0
pmdarima>=2.0.3

# Web Framework and API
fastapi>=0.100.0
uvicorn>=0.23.0
flask>=2.3.0
flask-cors>=4.0.0
flask-jwt-extended>=4.5.0

# Database
mysql-connector-python>=8.1.0
sqlalchemy>=2.0.0
alembic>=1.11.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
python-docx>=0.8.11
PyPDF2>=3.0.1

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Jupyter for notebooks
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
